rename module from codexsource to hawkapi, remove not used files and imports
