+++ SourceInterfaces/src/main/java/so/codex/sourceinterfaces/entity/AuthEntity.kt
@@ -1,9 +1,3 @@
 package so.codex.sourceinterfaces.entity
 
-data class AuthEntity(val email: String, val password: String) {
-    fun getMutation() = """mutation login(${"$"}email: String!, ${"$"}password: String!) {
-        |  login(email: ${"$"}email, password: ${"$"}password) {
-        |      accessToken    refreshToken  
-        |  }
-        |}""".trimMargin()
-}
+data class AuthEntity(val email: String, val password: String)
