--- SourceInterfaces/src/main/java/so/codex/sourceinterfaces/entity/RequestEntity.kt
@@ -1,3 +1,0 @@
-package so.codex.sourceinterfaces.entity
-
-data class RequestEntity(val query: String, val variables: Any)
