+++ SourceInterfaces/src/main/java/so/codex/sourceinterfaces/entity/SignUpEntity.kt
@@ -1,7 +1,3 @@
 package so.codex.sourceinterfaces.entity
 
-data class SignUpEntity(val email: String) {
-    fun getMutation() = """mutation signUp(${"$"}email: String!) {
-        |   signUp(email: ${"$"}email)
-        |}""".trimMargin()
-}
+data class SignUpEntity(val email: String)
