--- SourceInterfaces/src/main/java/so/codex/sourceinterfaces/response/CodexResponse.kt
@@ -1,3 +1,0 @@
-package so.codex.sourceinterfaces.response
-
-data class CodexResponse<T : Any?>(val data: T)
