+++ SourceInterfaces/src/main/java/so/codex/sourceinterfaces/response/SignUpResponse.kt
@@ -1,5 +1,3 @@
 package so.codex.sourceinterfaces.response
 
-data class SignUpResponse(
-    val signUp: Boolean
-)
+data class SignUpResponse(val signUp: Boolean)
