+++ SourceInterfaces/src/main/java/so/codex/sourceinterfaces/response/TokenResponse.kt
@@ -1,3 +1,3 @@
 package so.codex.sourceinterfaces.response
 
-class TokenResponse(val accessToken: String, val refreshToken: String)
+data class TokenResponse(val accessToken: String, val refreshToken: String)
