--- codexsource/src/main/AndroidManifest.xml
@@ -1,10 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="so.codex.codexsource" >
-    <application
-        android:allowBackup="true"
-        android:label="@string/app_name"
-        android:supportsRtl="true"
-        >
-
-    </application>
-</manifest>
