--- codexsource/src/main/java/so/codex/codexsource/SourceApi.kt
@@ -1,10 +1,0 @@
-package so.codex.codexsource
-
-import so.codex.sourceinterfaces.IAuthApi
-
-/**
- * Интерфейс, в котором определены все методы, которые возвращают соответсвутющие компоненты для взаимодействия с сервером
- */
-interface SourceApi {
-    fun getAuthApi(): IAuthApi
-}
