--- codexsource/src/main/java/so/codex/codexsource/exceptions/BaseHttpException.kt
@@ -1,3 +1,0 @@
-package so.codex.codexsource.exceptions
-
-open class BaseHttpException(message: String) : Exception(message)
