--- codexsource/src/main/java/so/codex/codexsource/exceptions/MultiHttpErrorsException.kt
@@ -1,3 +1,0 @@
-package so.codex.codexsource.exceptions
-
-class MultiHttpErrorsException(message: String, val errors: List<String>) : BaseHttpException(message)
