--- codexsource/src/main/java/so/codex/codexsource/exceptions/SomethingWentWrongException.kt
@@ -1,3 +1,0 @@
-package so.codex.codexsource.exceptions
-
-class SomethingWentWrongException : Exception("Something went wrong")
