+++ hawkapi/build.gradle
