+++ hawkapi/src/main/graphql/so/codex/hawkapi/SignUpMutation.graphql
