+++ hawkapi/src/main/graphql/so/codex/hawkapi/schema.json
