+++ hawkapi/src/main/java/so/codex/hawkapi/SourceApi.kt
@@ -1,0 +1,10 @@
+package so.codex.hawkapi
+
+import so.codex.sourceinterfaces.IAuthApi
+
+/**
+ * Интерфейс, в котором определены все методы, которые возвращают соответсвутющие компоненты для взаимодействия с сервером
+ */
+interface SourceApi {
+    fun getAuthApi(): IAuthApi
+}
