+++ hawkapi/src/main/java/so/codex/hawkapi/exceptions/BaseHttpException.kt
@@ -1,0 +1,3 @@
+package so.codex.hawkapi.exceptions
+
+open class BaseHttpException(message: String) : Exception(message)
