+++ hawkapi/src/main/java/so/codex/hawkapi/exceptions/MultiHttpErrorsException.kt
@@ -1,0 +1,3 @@
+package so.codex.hawkapi.exceptions
+
+class MultiHttpErrorsException(message: String, val errors: List<String>) : BaseHttpException(message)
