+++ hawkapi/src/main/java/so/codex/hawkapi/exceptions/SomethingWentWrongException.kt
@@ -1,0 +1,3 @@
+package so.codex.hawkapi.exceptions
+
+class SomethingWentWrongException : Exception("Something went wrong")
