+++ hawkapi/src/main/res/values/strings.xml
