+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':uicomponent', ':codexbl', ':codexsource', ':SourceInterfaces'
+include ':app', ':uicomponent', ':codexbl', ':hawkapi', ':SourceInterfaces'
